backend/requirements.txt
asgiref==3.8.1
attrs==24.3.0
dj-database-url==2.3.0
Django==5.1.5
django-heroku==0.3.1
django-redis==5.4.0
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
packaging==24.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.8.2
